requirements.txt
Django==3.2
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
psycopg2-binary==2.8.6
pandas==1.3.5
django-cors-headers==3.10.0